Fixed a bug in jump offsets and added labels for jumping.
Turns out that the disassembler tries to be smart and put the operand of jump
instructions as the 0 index byte from the beginning of the sequence which leads
to really confusing results. Igor-test now dumps the hex so you can actually
find out what the relative offset is to verify correctness.

I took out the compare generation for conditional jumps because I realized that
it would add an extra invisible predicate and it's size would not be accounted
for in the jump, so there needs to be a way of handling jumps that have gadgets
sequences that occur before them...