Generalizing how the input to statements are handled.
Parameters are now abstracted as a Param type. There are methods withParam,
withRegister, and withTempRegister that automatically handle loading any type of
Param into a temporary register and then freeing it after it is used. This
centralizes a lot of code that was floating around everywhere previously.

Still a WIP.