Optimized the shit out of forward jumps.
Modified the structure of Jump gadgets to include the size, which eliminates the
need to filtering by size. There is still some cruft code in Discover in load
and save to deal with the old style jump gadgets still in the library files :/.
But hey, 60x speed up in compilation time, whatever!